Move bench to benches folder
