+++ sheep/src/lib.rs
@@ -1,6 +1,3 @@
-#![feature(test)]
-extern crate test;
-
 #[cfg(feature = "amethyst")]
 extern crate serde;
