+++ sheep/src/pack/maxrects.rs
@@ -377,7 +377,6 @@ fn remove_redundant_rects(rects: &mut Vec<Rect>) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test::{self, Bencher};
 
     #[test]
     fn remove_redundant() {
@@ -443,27 +442,5 @@ mod tests {
             assert_eq!(bin.dimensions.0, 100);
             assert_eq!(bin.dimensions.1, 100);
         }
-    }
-
-    #[bench]
-    fn bench_pack(b: &mut Bencher) {
-        let mut sprites = (0..1000)
-            .map(|i| SpriteData::new(i, (100, 100)))
-            .collect::<Vec<SpriteData>>();
-
-        let smaller_sprites = (0..1000)
-            .map(|i| SpriteData::new(i, (80, 80)))
-            .collect::<Vec<SpriteData>>();
-
-        sprites.extend(smaller_sprites.into_iter());
-
-        let options = MaxrectsOptions::default()
-            .preferred_width(1000 * 100)
-            .preferred_height(1000 * 100);
-
-        b.iter(|| {
-            let result = MaxrectsPacker::pack(&sprites, options);
-            let _ = test::black_box(&result);
-        });
     }
 }
